% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny}
\alias{run_shiny}
\title{Run the shiny app
This app will show the temperature forecast based on the city and units during the next month}
\usage{
run_shiny()
}
\description{
Run the shiny app
This app will show the temperature forecast based on the city and units during the next month
}
